task 31: correction code style errors
